refactor testing in functor day2
